Created a Friend component
