+++ frontend/src/components/friends/FriendsList.css
